fix(windows): avoid narrowing conversion

+++ src/fastxs3d.cpp
@@ -107,7 +107,7 @@ auto projection(
 	const uint64_t pvoxels = psx * psx;
 
 	auto projectionfn = [&](auto dtype) {
-		auto cutout = py::array_t<decltype(dtype), py::array::f_style>(std::vector<ssize_t>{ 0, 0 });
+		auto cutout = py::array_t<decltype(dtype), py::array::f_style>(std::vector<int64_t>{ 0, 0 });
 
 		if (crop_distance == 0) {
 			return cutout.view(py::str(labels.dtype()));
@@ -132,7 +132,7 @@ auto projection(
 		bbox.x_max++;
 		bbox.y_max++;
 
-		cutout = py::array_t<decltype(dtype), py::array::f_style>(std::vector<ssize_t>{ bbox.sx(), bbox.sy() });
+		cutout = py::array_t<decltype(dtype), py::array::f_style>(std::vector<int64_t>{ bbox.sx(), bbox.sy() });
 	    auto cutout_ptr = reinterpret_cast<decltype(dtype)*>(cutout.request().ptr);
 
 	    int64_t csx = bbox.sx();
